Fix minmax index for nullable DateTime

+++ src/Storages/MergeTree/IMergeTreeDataPart.cpp
@@ -563,7 +563,7 @@ std::pair<DayNum, DayNum> IMergeTreeDataPart::getMinMaxDate() const
         if (hyperrectangle.left.isNull())
             return {};
 
-        return {DayNum(hyperrectangle.left.safeGet<UInt64>()), DayNum(hyperrectangle.right.safeGet<UInt64>())};
+        return {DayNum(static_cast<UInt16>(hyperrectangle.left.safeGet<UInt64>())), DayNum(static_cast<UInt16>(hyperrectangle.right.safeGet<UInt64>()))};
     }
     return {};
 }
